feat: Moved application card information into page, and resolved scrolling issues on chrome
